Changed float -> double and new resampling library

+++ src/analysis/filter/filter.cpp
@@ -1,11 +1,11 @@
 #include "filter.h"
 
-std::vector<float> Analysis::filter(const std::vector<float>& a, const std::vector<float>& x)
+std::vector<double> Analysis::filter(const std::vector<double>& a, const std::vector<double>& x)
 {
     const int nx = x.size();
     const int na = a.size();
 
-    std::vector<float> y(nx);
+    std::vector<double> y(nx);
 
     for (int i = 0; i < nx; ++i) {
         y[i] = 0.0f;
@@ -19,13 +19,13 @@ std::vector<float> Analysis::filter(const std::vector<float>& a, const std::vect
     return y;
 }
 
-std::vector<float> Analysis::filter(const std::vector<float>& b, const std::vector<float>& a, const std::vector<float>& x)
+std::vector<double> Analysis::filter(const std::vector<double>& b, const std::vector<double>& a, const std::vector<double>& x)
 {
     const int nx = x.size();
     const int nb = b.size();
     const int na = a.size();
 
-    std::vector<float> y(nx);
+    std::vector<double> y(nx);
 
     for (int i = 0; i < nx; ++i) {
         y[i] = 0.0f;
