+++ src/analysis/invglot/invglot.h
@@ -10,39 +10,39 @@
 namespace Analysis {
 
     struct InvglotResult {
-        float sampleRate;
-        std::vector<float> glotSig;
+        double sampleRate;
+        std::vector<double> glotSig;
     };
 
     class InvglotSolver {
     public:
         virtual ~InvglotSolver() {}
-        virtual InvglotResult solve(const float *x, int length, float sampleRate) = 0;
+        virtual InvglotResult solve(const double *x, int length, double sampleRate) = 0;
     };
 
     namespace Invglot {
         class IAIF : public InvglotSolver {
         public:
-            IAIF(float d);
-            InvglotResult solve(const float *x, int length, float sampleRate) override;
+            IAIF(double d);
+            InvglotResult solve(const double *x, int length, double sampleRate) override;
         private:
             std::unique_ptr<LinpredSolver> lpc;
-            float d;
+            double d;
         };
 
         class GFM_IAIF : public InvglotSolver {
         public:
-            GFM_IAIF(float d);
-            InvglotResult solve(const float *x, int length, float sampleRate) override;
+            GFM_IAIF(double d);
+            InvglotResult solve(const double *x, int length, double sampleRate) override;
         private:
             std::unique_ptr<LinpredSolver> lpc;
-            float d;
+            double d;
         };
 
         class AMGIF : public InvglotSolver {
         public:
             AMGIF(int J);
-            InvglotResult solve(const float *x, int length, float sampleRate) override;
+            InvglotResult solve(const double *x, int length, double sampleRate) override;
         private:
             void computeC();
             std::vector<Eigen::MatrixXd> C;
