+++ src/analysis/util/parabolic_interpolation.cpp
@@ -1,8 +1,8 @@
 #include "util.h"
 
-std::pair<float, float> Analysis::parabolicInterpolation(const std::vector<float>& array, int x) {
+std::pair<double, double> Analysis::parabolicInterpolation(const std::vector<double>& array, int x) {
     int x_adjusted;
-    float x_ = x;
+    double x_ = x;
 
     if (x < 1) {
         x_adjusted = (array[x] <= array[x + 1]) ? x : x + 1;
@@ -11,8 +11,8 @@ std::pair<float, float> Analysis::parabolicInterpolation(const std::vector<float
         x_adjusted = (array[x] <= array[x - 1]) ? x : x - 1;
     }
     else {
-        float den = array[x + 1] + array[x - 1] - 2 * array[x];
-        float delta = array[x - 1] - array[x + 1];
+        double den = array[x + 1] + array[x - 1] - 2 * array[x];
+        double delta = array[x - 1] - array[x + 1];
         return (den == 0) ? std::make_pair(x_, array[x])
                           : std::make_pair(x_ + delta / (2 * den),
                                   array[x] - delta * delta / (8 * den));
